Corrected Build Error / Moved types into ScriptingEngine_RegisterTypes
